refactor: extract skill candidate validation helpers
